Don’t dismiss unpresented view controllers
When using the Dropbox app to authorize access, the in-app auth view is not being presented. However, when we return to the app it was indiscriminately dismissing any presented view controllers it had. In the case of 1Password, this was a view controller we were presenting to setup Dropbox sync. Additionally, in certain other cases it was over-dismissing the presented view controller, resulting in more than one view controller being dismissed.