Prevent UI thread from blocking during batch upload
The `batchUploadFiles` call does not report back a response
until all uploads are completed. To do this, it calls sleep(1)
while it waits for (1) everything to finish, (2) a cancel request,
or (3) all uploads to finish.

The calls to sleep(1) are made inside a `setResponseBlock` handler
that is not passed a `queue` parameter. Because the queue isn't
specified, it runs on the main thread. This causes apps that call
this function to freeze on their main thread for a second.

This patch makes a dedicated NSOperationQueue for monitoring the
progress.